# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a modular payment switch API with inquiry, execute, and status check.",
        "title": "Mini Payment Switch API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8182",
    "basePath": "/",
    "paths": {
        "/api/v1/payments/execute": {
            "post": {
                "description": "Processes payment using an inquiry ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Execute Payment",
                "parameters": [
                    {
                        "description": "Execute Payment Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usecase.ExecutePaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/usecase.ExecutePaymentResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/payments/inquiry": {
            "post": {
                "description": "Validates account and calculates fees before payment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Payment Inquiry",
                "parameters": [
                    {
                        "description": "Inquiry Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usecase.InquiryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/usecase.InquiryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/payments/status/{trx_id}": {
            "get": {
                "description": "Retrieves the current status of a transaction",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Check Payment Status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction ID",
                        "name": "trx_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/usecase.CheckStatusResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "response.ErrorInfo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/response.ErrorInfo"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "usecase.CheckStatusResponse": {
            "type": "object",
            "properties": {
                "account_no": {
                    "type": "string"
                },
                "amount": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                },
                "trx_id": {
                    "type": "string"
                }
            }
        },
        "usecase.ExecutePaymentRequest": {
            "type": "object",
            "properties": {
                "inquiry_id": {
                    "type": "string"
                }
            }
        },
        "usecase.ExecutePaymentResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "trx_id": {
                    "type": "string"
                }
            }
        },
        "usecase.InquiryRequest": {
            "type": "object",
            "properties": {
                "account_no": {
                    "type": "string"
                },
                "amount": {
                    "type": "number"
                }
            }
        },
        "usecase.InquiryResponse": {
            "type": "object",
            "properties": {
                "account_no": {
                    "type": "string"
                },
                "admin_fee": {
                    "type": "number"
                },
                "amount": {
                    "type": "number"
                },
                "inquiry_id": {
                    "type": "string"
                },
                "total": {
                    "type": "number"
                }
            }
        }
    }
}